requirements.txt
annotated-types==0.5.0
anyio==3.7.1
cffi==1.15.1
click==8.1.7
cryptography==41.0.3
ecdsa==0.18.0
fastapi==0.101.1
greenlet==2.0.2
h11==0.14.0
idna==3.4
mypy-extensions==1.0.0
packaging==23.1
passlib==1.7.4
pathspec==0.11.2
platformdirs==3.10.0
pyasn1==0.5.0
pycparser==2.21
pydantic==2.2.1
pydantic_core==2.6.1
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.20
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.22.0
bcrypt==4.0.0
python-dotenv==1.0.0
pytest-asyncio==0.21.1
httpx==0.24.1
psycopg2-binary==2.9.7